A for loop was using an index variable of type byte, which would overflow for graphs with more than 127 levels - fixed.
git-svn-id: https://svn.cms.waikato.ac.nz/svn/weka/trunk@10502 e0a1b77d-ad91-4216-81b1-defd5f83fa92

+++ weka/src/main/java/weka/gui/graphvisualizer/HierarchicalBCEngine.java
@@ -513,7 +513,7 @@ public class HierarchicalBCEngine implements GraphConstants, LayoutEngine {
     // ****Assigning nodes to each level
     int levelsCounter[] = new int[maxLevel + 1];
     nodeLevels = new int[maxLevel + 1][];
-    for (byte i = 0; i < nodesLevel.length; i++) {
+    for (int i = 0; i < nodesLevel.length; i++) {
       if (nodeLevels[nodesLevel[i]] == null) {
         nodeLevels[nodesLevel[i]] = new int[levelCounts[nodesLevel[i]]];
       }
